Adding more logging, checking some null values, fixing Time mark save point

+++ src/main/java/com/extractor/as400/util/ConfigVerification.java
@@ -91,7 +91,7 @@ public class ConfigVerification {
         for (as400Iterator = serverStateList.iterator(); as400Iterator.hasNext();){
             ServerState server = as400Iterator.next();
             if (server.getServerDefAS400().getHostName().compareTo(serverToChange.getServerDefAS400().getHostName()) == 0
-            && server.getServerDefAS400().getServerId() == serverToChange.getServerDefAS400().getServerId()) {
+            && server.getServerDefAS400().getServerId().intValue() == serverToChange.getServerDefAS400().getServerId().intValue()) {
                 server.setStatus(newStatus);
             }
         }
@@ -102,7 +102,7 @@ public class ConfigVerification {
         for (as400Iterator = serverStateList.iterator(); as400Iterator.hasNext();){
             ServerState server = as400Iterator.next();
             if (server.getServerDefAS400().getHostName().compareTo(serverToChange.getServerDefAS400().getHostName()) == 0
-                    && server.getServerDefAS400().getServerId() == serverToChange.getServerDefAS400().getServerId()) {
+                    && server.getServerDefAS400().getServerId().intValue() == serverToChange.getServerDefAS400().getServerId().intValue()) {
                 return server;
             }
         }
@@ -116,7 +116,7 @@ public class ConfigVerification {
                 if (i!=j) {
                     ServerState search = serverStateList.get(j);
                     if (base.getServerDefAS400().getHostName().compareTo(search.getServerDefAS400().getHostName()) == 0
-                            && base.getServerDefAS400().getServerId() == search.getServerDefAS400().getServerId()) {
+                            && base.getServerDefAS400().getServerId().intValue() == search.getServerDefAS400().getServerId().intValue()) {
                         return true;
                     }
                 }
